refactor providerdata process
