+++ server/server/session/src/main/java/com/alipay/sofa/registry/server/session/bootstrap/SessionServerBootstrap.java
@@ -16,21 +16,6 @@
  */
 package com.alipay.sofa.registry.server.session.bootstrap;
 
-import java.lang.annotation.Annotation;
-import java.util.Collection;
-import java.util.Date;
-import java.util.Map;
-import java.util.concurrent.atomic.AtomicBoolean;
-
-import javax.annotation.Resource;
-import javax.ws.rs.Path;
-import javax.ws.rs.ext.Provider;
-
-import org.glassfish.jersey.server.ResourceConfig;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.ApplicationContext;
-import org.springframework.util.CollectionUtils;
-
 import com.alipay.sofa.registry.common.model.Node;
 import com.alipay.sofa.registry.common.model.constants.ValueConstants;
 import com.alipay.sofa.registry.common.model.metaserver.FetchProvideDataRequest;
@@ -51,11 +36,25 @@ import com.alipay.sofa.registry.server.session.node.NodeManager;
 import com.alipay.sofa.registry.server.session.node.NodeManagerFactory;
 import com.alipay.sofa.registry.server.session.node.RaftClientManager;
 import com.alipay.sofa.registry.server.session.node.SessionProcessIdGenerator;
+import com.alipay.sofa.registry.server.session.provideData.ProvideDataProcessor;
 import com.alipay.sofa.registry.server.session.registry.Registry;
 import com.alipay.sofa.registry.server.session.remoting.handler.AbstractClientHandler;
 import com.alipay.sofa.registry.server.session.remoting.handler.AbstractServerHandler;
 import com.alipay.sofa.registry.server.session.scheduler.ExecutorManager;
 import com.alipay.sofa.registry.task.batcher.TaskDispatchers;
+import org.glassfish.jersey.server.ResourceConfig;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.ApplicationContext;
+import org.springframework.util.CollectionUtils;
+
+import javax.annotation.Resource;
+import javax.ws.rs.Path;
+import javax.ws.rs.ext.Provider;
+import java.lang.annotation.Annotation;
+import java.util.Collection;
+import java.util.Date;
+import java.util.Map;
+import java.util.concurrent.atomic.AtomicBoolean;
 
 /**
  * The type Session server bootstrap.
@@ -108,6 +107,9 @@ public class SessionServerBootstrap {
 
     @Autowired
     private Registry                          sessionRegistry;
+
+    @Autowired
+    private ProvideDataProcessor              provideDataProcessorManager;
 
     private Server                            server;
 
@@ -309,20 +311,7 @@ public class SessionServerBootstrap {
         Object ret = sendMetaRequest(fetchProvideDataRequest, leaderUrl);
         if (ret instanceof ProvideData) {
             ProvideData provideData = (ProvideData) ret;
-            if (provideData.getProvideData() == null
-                || provideData.getProvideData().getObject() == null) {
-                LOGGER.info("Fetch session stop push switch no data existed,config not change!");
-                return;
-            }
-            String data = (String) provideData.getProvideData().getObject();
-            sessionServerConfig.setStopPushSwitch(Boolean.valueOf(data));
-            if (data != null) {
-                if (!Boolean.valueOf(data)) {
-                    //stop push init on,then begin fetch data schedule task
-                    sessionServerConfig.setBeginDataFetchTask(true);
-                }
-            }
-            LOGGER.info("Fetch session stop push data switch {} success!", data);
+            provideDataProcessorManager.fetchDataProcess(provideData);
         } else {
             LOGGER.info("Fetch session stop push switch data null,config not change!");
         }
@@ -334,16 +323,7 @@ public class SessionServerBootstrap {
         Object data = sendMetaRequest(fetchProvideDataRequest, leaderUrl);
         if (data instanceof ProvideData) {
             ProvideData provideData = (ProvideData) data;
-            if (provideData == null || provideData.getProvideData() == null
-                || provideData.getProvideData().getObject() == null) {
-                LOGGER
-                    .info("Fetch enableDataRenewSnapshot but no data existed, current config not change!");
-                return;
-            }
-            boolean enableDataRenewSnapshot = Boolean.parseBoolean((String) provideData
-                .getProvideData().getObject());
-            LOGGER.info("Fetch enableDataRenewSnapshot {} success!", enableDataRenewSnapshot);
-            this.sessionRegistry.setEnableDataRenewSnapshot(enableDataRenewSnapshot);
+            provideDataProcessorManager.fetchDataProcess(provideData);
         }
     }
