+++ server/server/session/src/main/java/com/alipay/sofa/registry/server/session/bootstrap/SessionServerConfiguration.java
@@ -68,6 +68,11 @@ import com.alipay.sofa.registry.server.session.node.service.DataNodeService;
 import com.alipay.sofa.registry.server.session.node.service.DataNodeServiceImpl;
 import com.alipay.sofa.registry.server.session.node.service.MetaNodeService;
 import com.alipay.sofa.registry.server.session.node.service.MetaNodeServiceImpl;
+import com.alipay.sofa.registry.server.session.provideData.ProvideDataProcessor;
+import com.alipay.sofa.registry.server.session.provideData.ProvideDataProcessorManager;
+import com.alipay.sofa.registry.server.session.provideData.processor.BlackListProvideDataProcessor;
+import com.alipay.sofa.registry.server.session.provideData.processor.RenewSnapshotProvideDataProcessor;
+import com.alipay.sofa.registry.server.session.provideData.processor.StopPushProvideDataProcessor;
 import com.alipay.sofa.registry.server.session.registry.Registry;
 import com.alipay.sofa.registry.server.session.registry.SessionRegistry;
 import com.alipay.sofa.registry.server.session.remoting.ClientNodeExchanger;
@@ -736,6 +741,39 @@ public class SessionServerConfiguration {
         @Bean
         public RenewService renewService() {
             return new DefaultRenewService();
+        }
+    }
+
+    @Configuration
+    public static class SessionProvideDataConfiguration {
+
+        @Bean
+        public ProvideDataProcessor provideDataProcessorManager() {
+            return new ProvideDataProcessorManager();
+        }
+
+        @Bean
+        public ProvideDataProcessor blackListProvideDataProcessor(ProvideDataProcessor provideDataProcessorManager) {
+            ProvideDataProcessor blackListProvideDataProcessor = new BlackListProvideDataProcessor();
+            ((ProvideDataProcessorManager) provideDataProcessorManager)
+                .addProvideDataProcessor(blackListProvideDataProcessor);
+            return blackListProvideDataProcessor;
+        }
+
+        @Bean
+        public ProvideDataProcessor renewSnapshotProvideDataProcessor(ProvideDataProcessor provideDataProcessorManager) {
+            ProvideDataProcessor renewSnapshotProvideDataProcessor = new RenewSnapshotProvideDataProcessor();
+            ((ProvideDataProcessorManager) provideDataProcessorManager)
+                .addProvideDataProcessor(renewSnapshotProvideDataProcessor);
+            return renewSnapshotProvideDataProcessor;
+        }
+
+        @Bean
+        public ProvideDataProcessor stopPushProvideDataProcessor(ProvideDataProcessor provideDataProcessorManager) {
+            ProvideDataProcessor stopPushProvideDataProcessor = new StopPushProvideDataProcessor();
+            ((ProvideDataProcessorManager) provideDataProcessorManager)
+                .addProvideDataProcessor(stopPushProvideDataProcessor);
+            return stopPushProvideDataProcessor;
         }
     }
 }
