+++ server/server/session/src/main/java/com/alipay/sofa/registry/server/session/scheduler/task/ProvideDataChangeFetchTask.java
@@ -16,24 +16,10 @@
  */
 package com.alipay.sofa.registry.server.session.scheduler.task;
 
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.HashMap;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Map;
-import java.util.Set;
-import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.TimeUnit;
-
-import org.springframework.util.CollectionUtils;
-
-import com.alipay.sofa.registry.common.model.constants.ValueConstants;
 import com.alipay.sofa.registry.common.model.metaserver.DataOperator;
 import com.alipay.sofa.registry.common.model.metaserver.NotifyProvideDataChange;
 import com.alipay.sofa.registry.common.model.metaserver.ProvideData;
 import com.alipay.sofa.registry.common.model.store.DataInfo;
-import com.alipay.sofa.registry.common.model.store.Subscriber;
 import com.alipay.sofa.registry.common.model.store.URL;
 import com.alipay.sofa.registry.common.model.store.Watcher;
 import com.alipay.sofa.registry.core.model.ReceivedConfigData;
@@ -45,16 +31,18 @@ import com.alipay.sofa.registry.remoting.Server;
 import com.alipay.sofa.registry.remoting.exchange.Exchange;
 import com.alipay.sofa.registry.server.session.bootstrap.SessionServerConfig;
 import com.alipay.sofa.registry.server.session.converter.ReceivedDataConverter;
-import com.alipay.sofa.registry.server.session.filter.blacklist.BlacklistConstants;
-import com.alipay.sofa.registry.server.session.filter.blacklist.BlacklistManager;
 import com.alipay.sofa.registry.server.session.node.service.MetaNodeService;
-import com.alipay.sofa.registry.server.session.registry.Registry;
-import com.alipay.sofa.registry.server.session.store.Interests;
-import com.alipay.sofa.registry.server.session.store.ReSubscribers;
+import com.alipay.sofa.registry.server.session.provideData.ProvideDataProcessor;
 import com.alipay.sofa.registry.server.session.store.Watchers;
 import com.alipay.sofa.registry.task.listener.TaskEvent;
 import com.alipay.sofa.registry.task.listener.TaskEvent.TaskType;
 import com.alipay.sofa.registry.task.listener.TaskListenerManager;
+
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.concurrent.ConcurrentHashMap;
 
 /**
  *
@@ -83,27 +71,21 @@ public class ProvideDataChangeFetchTask extends AbstractSessionTask {
 
     private final Exchange            boltExchange;
 
-    private final Interests           sessionInterests;
-
-    private final Registry            sessionRegistry;
-
-    private final BlacklistManager    blacklistManager;
-
     private NotifyProvideDataChange   notifyProvideDataChange;
+
+    private ProvideDataProcessor      provideDataProcessorManager;
 
     public ProvideDataChangeFetchTask(SessionServerConfig sessionServerConfig,
                                       TaskListenerManager taskListenerManager,
                                       MetaNodeService metaNodeService, Watchers sessionWatchers,
-                                      Exchange boltExchange, Interests sessionInterests,
-                                      Registry sessionRegistry, BlacklistManager blacklistManager) {
+                                      Exchange boltExchange,
+                                      ProvideDataProcessor provideDataProcessorManager) {
         this.sessionServerConfig = sessionServerConfig;
         this.taskListenerManager = taskListenerManager;
         this.metaNodeService = metaNodeService;
         this.sessionWatchers = sessionWatchers;
         this.boltExchange = boltExchange;
-        this.sessionInterests = sessionInterests;
-        this.sessionRegistry = sessionRegistry;
-        this.blacklistManager = blacklistManager;
+        this.provideDataProcessorManager = provideDataProcessorManager;
     }
 
     @Override
@@ -130,71 +112,12 @@ public class ProvideDataChangeFetchTask extends AbstractSessionTask {
         if (notifyProvideDataChange.getDataOperator() != DataOperator.REMOVE) {
             provideData = metaNodeService.fetchData(dataInfoId);
 
-            if (ValueConstants.STOP_PUSH_DATA_SWITCH_DATA_ID.equals(dataInfoId)) {
-                //push stop switch
-                if (provideData != null) {
-                    if (provideData.getProvideData() == null || provideData.getProvideData().getObject() == null) {
-                        LOGGER.info("Fetch session stop push switch no data existed,config not change!");
-                        return;
-                    }
-                    String data = (String) provideData.getProvideData().getObject();
-                    LOGGER.info("Fetch session stop push data switch {} success!", data);
-
-                    //receive stop push switch off
-                    if (data != null) {
-                        boolean switchData = Boolean.valueOf(data);
-                        boolean ifChange = sessionServerConfig.isStopPushSwitch() != switchData;
-                        sessionServerConfig.setStopPushSwitch(switchData);
-                        if (!switchData) {
-                            //avoid duplicate false receive
-                            if (ifChange) {
-                                fireReSubscriber();
-                            }
-                        } else {
-                            //stop push and stop fetch data task
-                            sessionServerConfig.setBeginDataFetchTask(false);
-                        }
-                    } else {
-                        LOGGER.error("Fetch session stop push data switch is null!");
-                    }
-                    return;
-                } else {
-                    LOGGER.info("Fetch session stop push switch data null,config not change!");
-                }
-                return;
-            } else if (ValueConstants.BLACK_LIST_DATA_ID.equals(dataInfoId)) {
-                //black list data
-                if (provideData.getProvideData() == null || provideData.getProvideData().getObject() == null) {
-                    LOGGER.info("Fetch session blacklist no data existed,current config not change!");
-                    return;
-                }
-                String data = (String) provideData.getProvideData().getObject();
-                if (data != null) {
-                    Map<String, Map<String, Set<String>>> blacklistConfigMap = blacklistManager
-                            .convertBlacklistConfig(data);
-                    clientOffBlackIp(blacklistConfigMap);
-                    LOGGER.info("Fetch session blacklist data switch {} success!", data);
-                } else {
-                    LOGGER.info("Fetch session blacklist data null,current config not change!");
-                }
-                return;
-            } else if (ValueConstants.ENABLE_DATA_RENEW_SNAPSHOT.equals(dataInfoId)) {
-                //stop renew switch
-                if (provideData == null || provideData.getProvideData() == null
-                    || provideData.getProvideData().getObject() == null) {
-                    LOGGER.info("Fetch enableDataRenewSnapshot but no data existed, current config not change!");
-                    return;
-                }
-                boolean enableDataRenewSnapshot =  Boolean.parseBoolean((String) provideData.getProvideData().getObject());
-                LOGGER.info("Fetch enableDataRenewSnapshot {} success!", enableDataRenewSnapshot);
-                this.sessionRegistry.setEnableDataRenewSnapshot(enableDataRenewSnapshot);
-                return;
-            }
-
             if (provideData == null) {
                 LOGGER.warn("Notify provider data Change request {} fetch no provider data!", notifyProvideDataChange);
                 return;
             }
+            provideDataProcessorManager.changeDataProcess(provideData);
+
         }
         DataInfo dataInfo = DataInfo.valueOf(dataInfoId);
 
@@ -228,54 +151,6 @@ public class ProvideDataChangeFetchTask extends AbstractSessionTask {
         }
     }
 
-    /**
-     * open push switch to push all reSubscribers
-     */
-    private void fireReSubscriber() {
-
-        //try catch avoid to error cancel beginDataFetchTask switch on
-        try {
-            //begin push fire data fetch task first,avoid reSubscriber push duplicate
-            sessionRegistry.fetchChangDataProcess();
-        } catch (Throwable e) {
-            LOGGER.error("Open push switch first fetch task execute error", e);
-        }
-
-        try {
-            //wait 1 MINUTES for dataFetch task evict duplicate subscriber push
-            TimeUnit.MINUTES.sleep(1);
-        } catch (InterruptedException e) {
-            LOGGER.error("Wait for dataFetch Task Interrupted!");
-        }
-
-        //fetch task process 1 minutes,can schedule execute fetch task
-        sessionServerConfig.setBeginDataFetchTask(true);
-
-        if (this.sessionInterests instanceof ReSubscribers) {
-            ReSubscribers reSubscriber = (ReSubscribers) sessionInterests;
-
-            Map<String/*dataInfoId*/, Map<String/*registerId*/, Subscriber>> reSubscribers = reSubscriber
-                    .getReSubscribers();
-
-            if (reSubscribers != null && !reSubscribers.isEmpty()) {
-                reSubscribers.forEach(
-                        (dataInfoId, subscribers) -> fireSubscriberMultiFetchTask(dataInfoId, subscribers.values()));
-                reSubscriber.clearReSubscribers();
-            }
-        }
-    }
-
-    private void fireSubscriberMultiFetchTask(String dataInfoId, Collection<Subscriber> subscribers) {
-        //trigger fetch data for subscriber,and push to client node
-        if (!CollectionUtils.isEmpty(subscribers)) {
-            TaskEvent taskEvent = new TaskEvent(dataInfoId, TaskType.SUBSCRIBER_MULTI_FETCH_TASK);
-            taskEvent.setAttribute(Constant.PUSH_CLIENT_SUBSCRIBERS, subscribers);
-            TASK_LOGGER.info("send " + taskEvent.getTaskType()
-                             + " subscribersSize:{},dataInfoId:{}", subscribers.size(), dataInfoId);
-            taskListenerManager.sendTaskEvent(taskEvent);
-        }
-    }
-
     private void firePushTask(ReceivedConfigData receivedConfigData, URL clientUrl) {
 
         Map<ReceivedConfigData, URL> parameter = new HashMap<>();
@@ -283,53 +158,6 @@ public class ProvideDataChangeFetchTask extends AbstractSessionTask {
         TaskEvent taskEvent = new TaskEvent(parameter, TaskType.RECEIVED_DATA_CONFIG_PUSH_TASK);
         TASK_LOGGER.info("send " + taskEvent.getTaskType() + " taskEvent:{}", taskEvent);
         taskListenerManager.sendTaskEvent(taskEvent);
-    }
-
-    private void clientOffBlackIp(Map<String, Map<String, Set<String>>> blacklistConfigMap) {
-
-        if (blacklistConfigMap != null) {
-            Set<String> ipSet = new HashSet();
-
-            for (Map.Entry<String, Map<String, Set<String>>> configEntry : blacklistConfigMap
-                .entrySet()) {
-                if (BlacklistConstants.FORBIDDEN_PUB.equals(configEntry.getKey())
-                    || BlacklistConstants.FORBIDDEN_SUB_BY_PREFIX.equals(configEntry.getKey())) {
-                    Map<String, Set<String>> typeMap = configEntry.getValue();
-                    if (typeMap != null) {
-                        for (Map.Entry<String, Set<String>> typeEntry : typeMap.entrySet()) {
-                            if (BlacklistConstants.IP_FULL.equals(typeEntry.getKey())) {
-                                if (typeEntry.getValue() != null) {
-                                    ipSet.addAll(typeEntry.getValue());
-                                }
-                            }
-                        }
-                    }
-                }
-
-            }
-
-            sessionRegistry.remove(getIpConnects(ipSet));
-        }
-    }
-
-    public List<String> getIpConnects(Set<String> _ipList) {
-
-        Server sessionServer = boltExchange.getServer(sessionServerConfig.getServerPort());
-
-        List<String> connections = new ArrayList<>();
-
-        if (sessionServer != null) {
-            Collection<Channel> channels = sessionServer.getChannels();
-            for (Channel channel : channels) {
-                String key = NetUtil.toAddressString(channel.getRemoteAddress());
-                String ip = key.substring(0, key.indexOf(":"));
-                if (_ipList.contains(ip)) {
-                    connections.add(key);
-                }
-            }
-        }
-
-        return connections;
     }
 
     private Map<String/*registerId*/, Watcher> getCache(String connectId) {
